% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRWGS.R
\name{compute_GP_allGeno}
\alias{compute_GP_allGeno}
\title{Run genomic prediction on all genotypes, output predicted values}
\usage{
compute_GP_allGeno(
  geno,
  pheno,
  traits,
  GP.method,
  runCV = FALSE,
  testSetGID = NULL,
  nreps = 10,
  nfolds = 10,
  h = 1,
  nb.mtry = 10,
  nIter = 6000,
  burnIn = 1000,
  ntree = 100,
  p2d.temp = NULL,
  nb.cores = 1,
  p2f.stats.cv = NULL,
  p2f.pred = NULL,
  verbose = 1
)
}
\arguments{
\item{geno}{genomic data with genotypes in row (GID in rownames) and marker in columns. Values should be column centered and scaled.}

\item{pheno}{phenotypic data with genotypes in row (in GID column) and traits in columns}

\item{traits}{character vector of trait names, must correspond to `pheno` column names}

\item{GP.method}{character vector of genomic prediction methods to use, must be one of "rrBLUP", "RKHS", "BayesA", "BayesB"}

\item{runCV}{logical, if TRUE, run cross-validation on the common genotypes in `pheno` and `geno`, default is FALSE}

\item{nreps}{number of repetitions for cross-validation, default is 10}

\item{nfolds}{number of folds for cross-validation, default is 10}

\item{h}{bandwith parameter for RKHS, default is 1.}

\item{nb.mtry}{number of randomly selected variables at each split for RandomForest, default is 10}

\item{nIter}{number of iterations for RKHS, default is 6000}

\item{burnIn}{number of burn-in iterations for RKHS, default is 1000}

\item{ntree}{number of trees for RandomForest, default is 100}

\item{p2d.temp}{path to directory to export temporary genomic prediction results, default is NULL (could cause error in parallelization if NULL).}

\item{nb.cores}{number of cores to parallelize the computation, default is 1 (no parallelization)}

\item{p2f.stats.cv}{path to file to export cross-validation genomic prediction results, default is NULL}

\item{p2f.pred}{path to file to export genotypic values from genomic prediction, default is NULL}

\item{verbose}{integer, level of verbosity, default is 1}

\item{testSetID}{GID of the test set genotypes, if NULL, will provide predicted genotypic values for all genotypes in `geno`}
}
\value{
a data frame with the predicted values for all genotypes and traits
}
\description{
Run genomic prediction on all genotypes, output predicted values
}
\seealso{
[compute_GP_methods()]
}
\author{
Charlotte Brault
}
