% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreedGSFunctions.R
\name{compute_GP_methods}
\alias{compute_GP_methods}
\title{Compute genomic prediction different methods}
\usage{
compute_GP_methods(
  geno,
  pheno,
  traits,
  GP.method,
  nreps = 10,
  nfolds = 10,
  h = 1,
  nb.mtry = 10,
  nIter = 6000,
  burnIn = 1000,
  ntree = 100,
  p2d.temp = NULL,
  nb.cores = 1,
  p2f.stats = NULL
)
}
\arguments{
\item{geno}{genomic data with genotypes in row (GID in rownames) and marker in columns. Values should be column centered and scaled.}

\item{pheno}{phenotypic data with genotypes in row (in GID column) and traits in columns. Phenotypic value should be corrected for year and location effects beforehand.}

\item{traits}{character vector of trait names}

\item{GP.method}{character vector of genomic prediction methods to use}

\item{nreps}{number of repetitions for cross-validation, default is 10}

\item{nfolds}{number of folds for cross-validation, default is 10}

\item{h}{bandwith parameter for RKHS, default is 1. If multiple values are provided, method will be RKHS Kernel Averaging.}

\item{nIter}{number of iterations for RKHS, default is 6000}

\item{burnIn}{number of burn-in iterations for RKHS, default is 1000}

\item{ntree}{number of trees for RandomForest, default is 100}

\item{p2d.temp}{path to directory to export temporary genomic prediction results, default is NULL (could cause error in parallelization if NULL).}

\item{nb.cores}{number of cores to parallelize the computation, default is 1 (no parallelization)}

\item{p2f.stats}{path to file to export genomic prediction results, default is NULL}
}
\description{
Compute genomic prediction different methods
}
