% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSWGS.R
\name{concordance_match_name}
\alias{concordance_match_name}
\title{Concordance of names between reference and match data
Matching of names will be done based on the spelling of the names, without punctuation.}
\usage{
concordance_match_name(
  match.name,
  ref.name = NULL,
  allowNoMatch = TRUE,
  corresp.list = NULL,
  verbose = 1
)
}
\arguments{
\item{match.name}{character vector of names to match (format)}

\item{ref.name}{character vector of reference names. Default is NULL but it needs to be provided if corresp.list is NULL.}

\item{allowNoMatch}{logical, keep all values from match.name even when there is no match, default is TRUE}

\item{corresp.list}{named list of correspondence between names, with correct spelling as name and possible synonym as character vector, default is NULL.
If provided, add new correspondence to the list. If there is a conflict between ref.names and corresp.list, the name in corresp.list will be used.}

\item{verbose}{integer, level of verbosity, default is 1}
}
\value{
a list of 3 elements: the correspondence data frame, the updated correspondence list, and a vector of new names to be applied to match name.
}
\description{
Concordance of names between reference and match data
Matching of names will be done based on the spelling of the names, without punctuation.
}
\author{
Charlotte Brault
}
