% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSWGS.R
\name{create_accessions_T3}
\alias{create_accessions_T3}
\title{Create table and file to add accession to T3}
\usage{
create_accessions_T3(
  dat = NULL,
  checkDB = TRUE,
  p2f = NULL,
  return_table = TRUE,
  accession_name = NULL,
  population_name = NULL,
  organization_name = NULL,
  synonym = NULL,
  variety = NULL,
  country_of_origin = NULL,
  notes = NULL,
  accession_number = NULL,
  purdy_pedigree = NULL,
  filial_generation = NULL,
  species_name = "Triticum aestivum"
)
}
\arguments{
\item{dat}{data frame with genotype information}

\item{checkDB}{logical, whether to check the T3 database with \code{BrAPI} package to exclude accessions already in the database, default is TRUE.}

\item{p2f}{path to file to write the output of the function}

\item{return_table}{logical, whether to return the table or not, default is TRUE.}

\item{accession_name}{character string, name of the column in \code{dat} that contains the accession name}

\item{population_name}{character string, name of the column in \code{dat} that contains the population name}

\item{organization_name}{charac6er string, name of the column in \code{dat} that contains the organization name}

\item{synonym}{character string, name of the column in \code{dat} that contains the synonym name}

\item{variety}{character string, name of the column in \code{dat} that contains the variety name}

\item{country_of_origin}{character string, name of the column in \code{dat} that contains the country of origin}

\item{notes}{character string, name of the column in \code{dat} that contains the notes}

\item{accession_number}{character string, name of the column in \code{dat} that contains the accession number}

\item{purdy_pedigree}{character string, name of the column in \code{dat} that contains the purdy pedigree}

\item{filial_generation}{character string, name of the column in \code{dat} that contains the filial generation}

\item{species_name}{character string, name of the species, default is "Triticum aestivum" (fixed).}
}
\value{
a data frame with the accessions to add to T3 if return_table is TRUE, otherwise a file is written to the path \code{p2f}
}
\description{
Create table and file to add accession to T3
}
\seealso{
\code{\link{create_trials_T3}}, \code{\link{create_phenot_T3}}
}
\author{
Charlotte Brault
}
