% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSWGS.R
\name{create_phenot}
\alias{create_phenot}
\title{Create table and file to add phenotypes in T3}
\usage{
create_phenot(
  dat = NULL,
  df_corresp_trait = NULL,
  p2f = NULL,
  return_table = TRUE,
  plot_name_col = NULL
)
}
\arguments{
\item{dat}{data frame with phenotypic data, containing at least the plot name and the traits}

\item{df_corresp_trait}{data frame with the correspondence between the column names in `dat` and the traits in T3, must countain the columns `T3_trait` and `corresp_col`.}

\item{p2f}{path to file to write the output of the function}

\item{return_table}{logical, whether to return the table or not, default is TRUE.}

\item{plot_name_col}{character string, name of the column in `dat` that contains the plot name, default is "plot_name"}
}
\value{
a data frame with the phenotypes to add to T3 if return_table is TRUE, otherwise a file is written to the path `p2f`.
}
\description{
Create table and file to add phenotypes in T3
}
