% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSWGS.R
\name{create_phenot_T3}
\alias{create_phenot_T3}
\title{Create table and file to add phenotype data in T3}
\usage{
create_phenot_T3(
  dat = NULL,
  df_corresp_trait = NULL,
  p2f = NULL,
  return_table = TRUE,
  plot_name_col = NULL
)
}
\arguments{
\item{dat}{data frame with phenotypic data, containing at least the plot name and the traits}

\item{df_corresp_trait}{data frame with the correspondence between the column names in \code{dat} and the traits in T3, must countain the columns \code{T3_trait} and \code{corresp_col}.}

\item{p2f}{path to file to write the output of the function}

\item{return_table}{logical, whether to return the table or not, default is TRUE.}

\item{plot_name_col}{character string, name of the column in \code{dat} that contains the plot name, default is "plot_name"}
}
\value{
a data frame with the phenotypes to add to T3 if return_table is TRUE, otherwise a file is written to the path \code{p2f}.
}
\description{
Create table and file to add phenotype data in T3
}
\seealso{
\code{\link{create_accessions_T3()}}, \code{\link{create_trials_T3()}}
}
