% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRWGS.R
\name{format_curate_vcf}
\alias{format_curate_vcf}
\title{Format and curate vcf file}
\usage{
format_curate_vcf(
  vcf.p2f = NULL,
  matrix.gt = NULL,
  mrk.info = NULL,
  corresp.geno.name = NULL,
  p2f.export.vcf = NULL,
  IDnum = FALSE,
  remove.chrUkn = TRUE,
  check.mrk.dups = TRUE,
  remove.nonPolyMrk = TRUE,
  tresh.heterozygous = NULL,
  check.geno.dups = TRUE,
  thresh.NA.ind = 0.5,
  thresh.NA.mrk = 0.5,
  format.names = FALSE,
  concordance.function = FALSE,
  corresp.geno.list = NULL,
  imputation = NULL,
  p2f.beagle = NULL,
  thresh.MAF = NULL,
  verbose = 1
)
}
\arguments{
\item{vcf.p2f}{path to the vcf file}

\item{matrix.gt}{matrix of genotypes, must contains CHROM and POS columns, default is NULL}

\item{mrk.info}{data frame with marker information pulled from the vcf, with columns CHROM, POS, ID, REF, ALT, QUAL, FILTER, INFO, FORMAT, default is NULL. Needed if `vcf.p2f` is not provided and `p2f.export.vcf` is not null.}

\item{corresp.geno.name}{table of correspondence between genotype names, with the first column being the name to be matched and the second one being the updated name.}

\item{p2f.export.vcf}{path to export curated vcf file, default is NULL (no export), useful for further Beagle imputation.}

\item{IDnum}{logical, remove the prefix idxx. from the genotype name, default is FALSE}

\item{remove.chrUkn}{logical, remove markers from unknown chromosome, default is TRUE}

\item{check.mrk.dups}{logical, check for duplicated markers, default is TRUE}

\item{remove.nonPolyMrk}{logical, remove non-segregating markers, default is TRUE}

\item{tresh.heterozygous}{numeric, threshold of the maximum proportion of heterozygous genotypes, default is NULL}

\item{check.geno.dups}{logical, check for duplicated genotypes, default is TRUE}

\item{thresh.NA.ind}{numeric, threshold of the maximum proportion missing values for individuals, default is 0.5}

\item{thresh.NA.mrk}{numeric, threshold of the maximum proportion of missing values for markers, default is 0.5}

\item{format.names}{logical, use a custom function to transform genotype names, default is FALSE}

\item{imputation}{character, impute missing values with kNNI or Beagle or don't impute, default is NULL (no imputation)}

\item{p2f.beagle}{path to export file for Beagle imputation, default is NULL}

\item{thresh.MAF}{numeric, threshold of minor allele frequency, default is NULL}

\item{verbose}{numeric, level of verbosity, default is 1}
}
\value{
data frame with 0/1/2 values from the curated vcf file, with genotypes in row and markers in columns.
}
\description{
Format and curate vcf file
}
\author{
Charlotte Brault
}
