% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSWGS.R
\name{format_names}
\alias{format_names}
\title{Format accession names}
\usage{
format_names(
  names,
  toupper = TRUE,
  dashToUnderscore = FALSE,
  pedigree = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{names}{character vector of accession names}

\item{toupper}{logical, set to upper case, default is TRUE}

\item{dashToUnderscore}{logical, transform dash to underscore, default is FALSE}

\item{pedigree}{is character vector a pedigree? if TRUE, don't remove slash strings}

\item{verbose}{logical, default to FALSE}
}
\value{
list of initial and corrected names
}
\description{
Format accession names
}
