% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRWGS.R
\name{format_phenot}
\alias{format_phenot}
\title{Format phenotypic data from GrainGenes (Excel tables)}
\usage{
format_phenot(p2d, years, locs, traits, cols2rem = NULL, distMatchTrait = 8)
}
\arguments{
\item{p2d}{path to directory where tables are saved}

\item{years}{numeric vector of years to look for}

\item{locs}{character vector of tab names to look for (including Entry) or to location names to identify the trial.
If several names are corresponding to one trial, repeat the different versions in the vector and add the final name as vector name.}

\item{traits}{character vector of trait names to look for.
If several names are corresponding to one trait, name the different versions and use as vector name the sought version.
for example: traits=c("VSK","Heading","FDK"); names(traits)=c("VSK","HD","VSK")}

\item{cols2rem}{character vector of column names to remove, to avoid bad matching}

\item{distMatchTrait}{numeric value, distance for string matching. Default is 8.
Increased distance would lead to more matching and is more prone to errors.}
}
\value{
list of 4 components:
- var.match.info: data frame of variable matching
- sheet.match.info: data frame of sheet matching (finding the relevant tabs)
- phenot: data frame of combined phenotypic data for all years and locations
- entry.info: data frame of combined genotype information
}
\description{
Format phenotypic data from GrainGenes (Excel tables)
}
\examples{
# p2d <- here("data","URSN_tables")
out <- format_phenot(p2d=p2d, years=year_ranges, locs=locs2check,
                     traits=traits2check,
                      cols2rem=c('DISK',"VRI",
                                 "VIBE FDK *","VIBE_FDK"))

}
\author{
Charlotte Brault
}
