% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRWGS.R
\name{getFolds}
\alias{getFolds}
\title{Get cross-validation folds}
\usage{
getFolds(index, nb.folds = 5, seed = NULL)
}
\arguments{
\item{index}{length of total set to select from, numeric vector of length 1}

\item{nb.folds}{number of fold to divide, default is 5}

\item{seed}{(optional) for reproducibility, set a seed.}
}
\value{
list of testing set indexes for each fold.
}
\description{
Get cross-validation folds
}
\author{
Charlotte Brault
}
