% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRWGS.R
\name{getGenoTas_to_DF}
\alias{getGenoTas_to_DF}
\title{Convert TASSEL object to genotypic data frame}
\usage{
getGenoTas_to_DF(tasGeno)
}
\arguments{
\item{tasGeno}{Genotypic object from TASSEL}
}
\value{
matrix of genotypic data in 0/1/2 data frame format
}
\description{
Convert TASSEL object to genotypic data frame
}
\author{
Vishnu Ramasubramanian
}
\keyword{internal}
