% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSWGS.R
\name{GGAIN}
\alias{GGAIN}
\title{Estimate genetic gain for selected traits}
\usage{
GGAIN(data, traits, first_year)
}
\arguments{
\item{data}{data frame with columns for traits and first year of evaluation}

\item{traits}{character vector of trait names}

\item{first_year}{character vector of column names for first year of evaluation.
The first year will be subtracted by the minimal first year for the estimation of the intercept and percentage of change}
}
\value{
data frame with columns for trait, intercept, slope, R2.adj, pvalue and percentage of change
}
\description{
Estimate genetic gain for selected traits
}
\author{
Charlotte Brault
}
