% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSWGS-package.R
\docType{package}
\name{HRSWGS-package}
\alias{HRSWGS}
\alias{HRSWGS-package}
\title{HRSWGS: Hard Red Spring Wheat breeding Genomic Selection toolbox}
\description{
Custom function to help data wrangling and formatting for using genomic selection in hard red spring wheat breeding programs
}
\seealso{
Useful links:
\itemize{
  \item \url{https://chabrault.github.io/HRSWGS/}
}

}
\author{
\strong{Maintainer}: Charlotte Brault \email{charlotte.brault@live.com}

}
\keyword{internal}
