% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSWGS.R
\name{pal_21wheatchr}
\alias{pal_21wheatchr}
\title{Color palette for the 21 wheat chromosomes}
\usage{
pal_21wheatchr()
}
\value{
a vector of colors for the 21 wheat chromosomes
}
\description{
Color palette for the 21 wheat chromosomes
}
\examples{
palette=pal_21wheatchr()[c(1:21)]
}
