% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSWGS.R
\name{pal_loc}
\alias{pal_loc}
\title{Color palette for locations}
\usage{
pal_loc()
}
\value{
a vector of colors for the locations
}
\description{
Color palette for locations
}
\examples{
palette=pal_loc()[c("GLL","MRD","LGD","CRK","CRT","PRP","FRG","BRK","STP")]
}
