% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSWGS.R
\name{pal_orga}
\alias{pal_orga}
\title{Color palette for breeding programs}
\usage{
pal_orga()
}
\value{
a vector of colors for the breeding programs
}
\description{
Color palette for breeding programs
}
\examples{
palette=pal_orga()[c("UMN","NDSU","SDSU","MSU","Check","Unknown")]
}
