% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSWGS.R
\name{pal_rand}
\alias{pal_rand}
\title{Color palette for random colors}
\usage{
pal_rand(n = 25)
}
\arguments{
\item{n}{number of colors to return, default is 25}
}
\value{
a vector of colors
}
\description{
Color palette for random colors
}
\examples{
palette=pal_rand(n=5)
}
