% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSWGS.R
\name{print_table}
\alias{print_table}
\title{Print nice table using DT}
\usage{
print_table(table, rownames = FALSE, digits = 3, ...)
}
\arguments{
\item{table}{data frame to print}

\item{rownames}{logical, if TRUE, print row names, default is FALSE}

\item{digits}{integer, number of digits to print for numeric columns, default is 3}

\item{...}{additional arguments to pass to DT::datatable}
}
\value{
a DT datatable object
}
\description{
Print nice table using DT
}
\examples{
print_table(mtcars)
}
